requirements.txt
asgiref==3.8.1
cx-Oracle==8.3.0
Django==3.2.25
django-cors-headers==3.14.0
djangorestframework==3.14.0
drf-yasg==1.20.0
inflection==0.5.1
packaging==25.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1